refactor unitary calculations (#355)
