+++ qbraid/interface/__init__.py
@@ -34,10 +34,10 @@ from .calculate_unitary import (
     UnitaryCalculationError,
     circuits_allclose,
     random_unitary_matrix,
+    rev_qubits_unitary,
     to_unitary,
     unitary_to_little_endian,
 )
 from .convert_to_contiguous import ContiguousConversionError, convert_to_contiguous
 from .draw import circuit_drawer
 from .programs import random_circuit
-from .qbraid_pyquil.tools import rev_qubits_unitary
